requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
datasets==2.21.0
dill==0.3.8
distlib==0.3.8
fastapi==0.88.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
huggingface-hub==0.24.6
identify==2.6.0
idna==3.8
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
nodeenv==1.9.1
numpy==2.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
psycopg2-binary==2.9.9
py==1.11.0
pyarrow==17.0.0
pycparser==2.22
pydantic==1.10.18
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.7
starlette==0.22.0
sympy==1.13.2
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.20.0
virtualenv==20.26.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.9.7
